Start talking about variables and data types
Editor's Note: This isn't done yet! I started this too late (It's 1:15 right now), but I'll try to finish it later. Feel free to change the explanations or edit what's necessary. what's clear to me might not be clear to other people. Also, if this is a course for people of all skill levels, I'm debating on whether or not to put a section dedicated to explaining what exactly programming is. Not sure if this will fit or not though, it's up to you